Potential fix for code scanning alert no. 5: Multiplication result converted to larger type

+++ src/core/qrcodegen.cpp
@@ -655,7 +655,7 @@ long QrCode::getPenaltyScore() const {
 	// Compute the smallest integer k >= 0 such that (45-5k)% <= dark/total <= (55+5k)%
 	int k = static_cast<int>((std::abs(dark * 20L - total * 10L) + total - 1) / total) - 1;
 	assert(0 <= k && k <= 9);
-	result += k * PENALTY_N4;
+	result += static_cast<long>(k) * PENALTY_N4;
 	assert(0 <= result && result <= 2568888L);  // Non-tight upper bound based on default values of PENALTY_N1, ..., N4
 	return result;
 }
